requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altered-midas==0.0.1
appdirs==1.4.4
aria2p==0.12.0
asciimatics==1.15.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
boosted-depth==0.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
chrislib==0.0.1
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
fastcore==1.7.12
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gdown==5.2.0
grpcio==1.66.2
huggingface-hub==0.25.2
idna==3.10
imageio==2.35.1
intrinsic==0.0.1
intrinsic-compositing==0.0.1
Jinja2==3.1.4
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
lightning-utilities==0.11.7
loguru==0.7.2
Markdown==3.7
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
omnidata-tools==0.0.21
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
propcache==0.2.0
protobuf==5.28.2
pyfiglet==1.0.2
pyparsing==3.1.4
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
soupsieve==2.6
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-logger==0.1.0
tifffile==2024.9.20
timm==1.0.9
tomli==2.0.2
torch==2.4.1
torchmetrics==1.4.2
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.4
yarl==1.14.0
